+++ dev/v1.OpenTranslationEngine.php
@@ -7,107 +7,6 @@ use Attogram\Router\Router;
 
 class V1OpenTranslationEngine
 {
-    const OTE_NAME    = 'Open Translation Engine';
-    const OTE_VERSION = '2.0.0-alpha.1';
-
-    /** @var Router */
-    private $router;
-
-    public function route()
-    {
-        $this->router = new Router;
-        $this->router->setForceSlash(true);
-        $this->setRoutes();
-        $control = $this->router->match();
-        if (!$control) {
-            $this->pageNotFound();
-
-            return;
-        }
-        /** @var string $control */
-        $this->pageHeader();
-        $this->includeTemplate($control);
-        $this->pageFooter();
-    }
-
-    private function setRoutes()
-    {
-        // Public Routes
-        $this->router->allow('/',            'home');
-        $this->router->allow('/languages/',  'languages');
-        $this->router->allow('/dictionary/', 'dictionary');
-        $this->router->allow('/word/',       'word');
-        $this->router->allow('/slush_pile/', 'slush_pile');
-        $this->router->allow('/search/',     'search');
-        $this->router->allow('/export/',     'export');
-        $this->router->allow('/history/',    'history');
-        $this->router->allow('/readme/',     'readme');
-
-        // Admin Routes
-        $this->router->allow('/import/',          'import');
-        $this->router->allow('/languages-admin/', 'languages-admin');
-        $this->router->allow('/tags-admin/',      'tags-admin');
-        $this->router->allow('/events/',          'events');
-        $this->router->allow('/info/',            'info');
-        $this->router->allow('/db-admin/',        'db-admin');
-        $this->router->allow('/db-tables/',       'db-tables');
-        $this->router->allow('/find_3rd_level/',  'find_3rd_level');
-        $this->router->allow('/check.php',        'check.php');
-    }
-
-    /**
-     * @param string $message
-     */
-    private function pageNotFound(string $message = '404 Page Not Found')
-    {
-        header('HTTP/1.0 404 Not Found', true, 404);
-        $this->pageHeader();
-        print $message;
-        $this->pageFooter();
-    }
-
-    /**
-     * @param string $message
-     */
-    private function controlNotFound(string $message = '500 Control Not Found')
-    {
-        header('HTTP/1.0 500 Internal Server Error', true, 500);
-        $this->pageHeader();
-        print $message;
-        $this->pageFooter();
-    }
-
-    /**
-     * @param string $title
-     */
-    public function pageHeader(string $title = '')
-    {
-        if (!$title) {
-            $title = OpenTranslationEngine::OTE_NAME . ' v' . OpenTranslationEngine::OTE_VERSION;
-        }
-        $this->includeTemplate('header');
-    }
-
-    public function pageFooter()
-    {
-        $this->includeTemplate('footer');
-    }
-
-    /**
-     * @param string $control
-     */
-    private function includeTemplate(string $control)
-    {
-        $template = __DIR__ . '/../template/' . $control . '.php';
-        if (!file_exists($template) || !is_readable($template)) {
-            print 'Template [' . $control . '] Not Found';
-            return;
-        }
-
-        /** @noinspection PhpIncludeInspection */
-        include $template;
-    }
-
     // v1 ----------------------------------------------
 
     public $path = './';
